ajax requests
split off take survey from make survey
