+++ assets/scripts/takesurvey/events.js
@@ -1,0 +1,7 @@
+'use strict';
+//
+const getFormFields = require('../../../lib/get-form-fields');
+
+const api = require('./api');
+const ui = require('./ui');
+const index = require('../index.js');
