+++ assets/scripts/takesurvey/ui.js
@@ -1,0 +1,3 @@
+'use strict';
+
+const app = require('../app.js');
